Fixed MSVC warning
git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/branches/v3_12@30675 023b2edf-31b2-4de3-b41e-bca80c47788f

+++ include/opal/mediastrm.h
@@ -432,7 +432,7 @@ class OpalMediaStream : public PSafeObject
 
     OpalConnection & connection;
     unsigned         sessionID;
-    unsigned         m_sequenceNumber;
+    WORD             m_sequenceNumber;
     PString          identifier;
     OpalMediaFormat  mediaFormat;
     bool             m_paused;
